Java layer revisited

+++ src/java/lang/Math.h
@@ -15,7 +15,7 @@ public:
     static long round(double a);
     static double log(double a);
     static float log(float a);
-    static float log10(float a);
+    static double log10(double a);
     static double sin(double a);
     static float sin(float a);
     static double cos(double a);
@@ -73,8 +73,8 @@ Math::log(double a) {
     return std::log(a);
 }
 
-inline float
-Math::log10(float a) {
+inline double
+Math::log10(double a) {
     return std::log10(a);
 }
